requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
